+++ package.json
@@ -58,5 +58,6 @@
     "prettier": "^2.5.1",
     "prettier-eslint": "^13.0.0",
     "react-app-rewired": "^2.1.11"
-  }
+  },
+  "proxy": "https://storage-fe.fastraffic.io"
 }
