+++ src/constants/constants.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = 'localhost';
